+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **[NEW]** Add `id` prop to `TheVoice`
 - **[NEW]** Add focus styles to `ItemRadio` (a11y)
 - **[NEW]** Add `ariaLabelledBy` prop to `ItemRadioGroup` (a11y)
+- **[FIX]** Enhance a11y by linking the content id with aria attributes on `SuccessModal`
 - [...]
 
 # v15.0.4 (20/11/2019)